integrate reset steps within run function. Remove reset function
